Fix compiler warnings and make warnings fatal (#5375)

+++ src/ray/raylet/node_manager.cc
@@ -1057,7 +1057,7 @@ void NodeManager::ProcessFetchOrReconstructMessage(
     const std::shared_ptr<LocalClientConnection> &client, const uint8_t *message_data) {
   auto message = flatbuffers::GetRoot<protocol::FetchOrReconstruct>(message_data);
   std::vector<ObjectID> required_object_ids;
-  for (size_t i = 0; i < message->object_ids()->size(); ++i) {
+  for (int64_t i = 0; i < message->object_ids()->size(); ++i) {
     ObjectID object_id = from_flatbuf<ObjectID>(*message->object_ids()->Get(i));
     if (message->fetch_only()) {
       // If only a fetch is required, then do not subscribe to the
@@ -1505,7 +1505,7 @@ void NodeManager::SubmitTask(const Task &task, const Lineage &uncommitted_lineag
           // The actor is local.
           int64_t expected_task_counter = GetExpectedTaskCounter(
               actor_registry_, spec.ActorId(), spec.ActorHandleId());
-          if (spec.ActorCounter() < expected_task_counter) {
+          if (static_cast<int64_t>(spec.ActorCounter()) < expected_task_counter) {
             // A task that has already been executed before has been found. The
             // task will be treated as failed if at least one of the task's
             // return values have been evicted, to prevent the application from
@@ -1724,7 +1724,7 @@ bool NodeManager::AssignTask(const Task &task) {
     // expected task counter.
     int64_t expected_task_counter =
         GetExpectedTaskCounter(actor_registry_, spec.ActorId(), spec.ActorHandleId());
-    RAY_CHECK(spec.ActorCounter() == expected_task_counter)
+    RAY_CHECK(static_cast<int64_t>(spec.ActorCounter()) == expected_task_counter)
         << "Expected actor counter: " << expected_task_counter << ", task "
         << spec.TaskId() << " has: " << spec.ActorCounter();
   }
@@ -2309,7 +2309,7 @@ void NodeManager::ForwardTask(
         // Iterate through the object's arguments. NOTE(swang): We do not include
         // the execution dependencies here since those cannot be transferred
         // between nodes.
-        for (int i = 0; i < spec.NumArgs(); ++i) {
+        for (size_t i = 0; i < spec.NumArgs(); ++i) {
           int count = spec.ArgIdCount(i);
           for (int j = 0; j < count; j++) {
             ObjectID argument_id = spec.ArgId(i, j);
